百分之85

+++ src/main/java/com/pufaschool/conn/filter/JwtAuthenticationTokenFilter.java
@@ -68,7 +68,7 @@ public class JwtAuthenticationTokenFilter extends OncePerRequestFilter {
         }
 
         //取出userId
-        Long id = Long.valueOf((Integer) claims.get("userId"));
+        Long id = Long.valueOf((Long) claims.get("userId"));
 
         //取出username
         String username = (String) claims.get("username");
